pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "viniihds", email = "viniciushsilva2210@gmail.com"},
]
dependencies = [
    "django>=4.2.4",
    "djangorestframework>=3.14.0",
    "django-extensions>=3.2.3",
    "pydotplus>=2.0.2",
    "setuptools>=68.1.2",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "isort>=5.12.0",
    "black>=23.9.1",
]

[tool.black]
line-length = 120